tests/cpydiff/modules_random_getrandbits.py: Removed example code. examples/getmorebits.py: Moved the example function over here.
I've removed the example function from within the documentaton and instead
I've moved it to the examples folder. I don't know the best way to link to
a folder from within the docs so I simply reference it from the workaround
message that's given. It's the same function but now fixed since it had a
bug in it originally where it could try to get 0 random bits which would
cause an error.