extmod/modurandom.c: getrandombits() method has proper error message. tests/cpydiff: Random module differences for getrandbits and randint. tests/cpydiff: Int type describes lack of bit length. examples/getmorebits.py: Example function to get arbritrary number of bits.
extmod/modurandom.c
The random module's getrandbits() method didn't give a proper error
message when calling it with a value that was outside of the range
of 1-32. Now instead of simply giving a ValueError it gives an error
message that states that the number of bits must be between 1 and 32.

tests/cpydiff:modules_random_getrandbits.py;modules_random_randint.py
Since the random module's methods getrandbits() and randint() differe
from the CPython random module's version of the methods tests have been
added. For getrandbits the relevant documentation is shown and added to
the docs. The same is given for randint method so that the information
is more easily found.

examples/getmorebits.py
This function is the example function referenced in the
modules_random_getrandbits.py workaround. It allows getting an arbitrary
number of random bits.

tests/cpydiff/types_int_bit_length.py
Finally since the Int object lacks the bit_length() method ther is a test
for that method also to include within the docs. It now shows the person
what they can see on micropython compared to CPython.