examples/getmorebits.py: Changd to while loop, remaining bits bux fixed.
The original code was a while loop, so it's been changed to the more
pythonic for in loop. Also before I wasn't making sure that the number
of bits that we were asking for was actually more than 0. As such an
error would have been caused if the person was asking for an even multiple
of 32 bits. This error has been fixed and is now bug free(I hope).