examples/getmorebits.py:Comparison was wrong, and method wasn't imported.
If the code was as described then it'd never run. I forgot to just import
getrandbits method from random module. Further I was comparing with the
number of bits instead of the remainder. This would still cause the error.
Thus it'd never run. This commit fixes that issue and will make the code
actually runable even if the number is an even multiple of 32.