tests: Change default Python command used on Windows.
Default to just calling python since that is most commonly
available: the official installer or zipfiles from python.org,
anaconda, nupkg all result in python being available but not
python3.  In other words: the default used so far is wrong.
Note that os.name is 'posix' when running the python version
which comes with Cygwin or MSys2 so they are not affected by
this.  However of all possible ways to get Python on Windows,
only Cygwin provides no python command so update the default
way for running tests in the README.