update uzlib to v2.9.5
uzlib isn't actually used in any firmwares, but is built into the
"unix" port used for testing.

The main benefit of the update is to fix problems encountered on
Windows, as the old ref of uzlib had filenames with embedded colons;
this has been fixed upstream.

uzlib seems to have been reabsed since the version that we took; this
doesn't really matter to us.