tools: Remove obsolete upip bootstrap script.
The upip module is frozen into ports supporting it, and it is included
in the source tree, so there is no need to get it from PyPi.  Moreover
the PyPi package referred to is an out-of-date version of upip which is
basically unrelated our upip.py because the source is taken from a fork
of micropython-lib instead of this repository.