ports/rp2: Add JMP PIN support for PIO.
PIO state machines can make a conditional jump on the state of a pin:
the `JMP PIN` command. This requires the pin to be configured with
`sm_config_set_jmp_pin`, but until now we didn't have a way of doing
that in MicroPython.

This patch adds a new `jmp_pin=None` argument to `StateMachine`. If it
is not `None` then we try to interpret it as a Pin, and pass its value
to `sm_config_set_jmp_pin`.

This change allows for porting the `uart_rx` example to MicroPython.
This example uses `jmp(pin)`, so without the `jmp_pin` argument it
doesn't work: the branch to `good_stop` is never taken, so nothing is
pushed into the FIFO.

```py
from machine import Pin, UART

rx = Pin(5, Pin.IN, Pin.PULL_UP)

@rp2.asm_pio(
    in_shiftdir=rp2.PIO.SHIFT_RIGHT,
)
def uart_rx():
    wrap_target()
    label("start")
    wait(0, pin, 0)
    set(x, 7)                        [10]
    label("bitloop")
    in_(pins, 1)
    jmp(x_dec, "bitloop")                  [6]
    jmp(pin, "good_stop")
    irq(block, 4)
    wait(1, pin, 0)
    jmp("start")
    label("good_stop")
    push(block)
    wrap()

baud = 9600

sm = rp2.StateMachine(0, uart_rx, freq=baud*8, in_base=rx, jmp_pin=rx)
sm.active(1)

while True:
    # Print hex bytes to the console.
    b = sm.get() >> 24
    print("{:02x}".format(b), end=" ")
```

Signed-off-by: Tim Radvan <tim@tjvr.org>