esp32/uart: Add flow keyword argument to enable hardware flow control.
This enables optional support for the hardware UART to use the RTS
and/or CTS pins for flow control.

The new "flow" constructor keyword specifies a bitmask of RTS and/or
CTS. This matches the interface used by machine.UART on STM32.

Previously on ESP32 it was possible to specify which pins to use for the
RTS and CTS signals, but hardware flow control was never functional: CTS
was not checked before transmitting bytes, and RTS was always driven
high (signalling no buffer space available). With this patch, CTS and
RTS both operate as expected.

This also includes an update to the machine.UART documentation.

Signed-off-by: Will Sowerbutts <will@sowerbutts.com>