rp2_pio.c: Add 3 state machine helper functions.
statemachine.restart: Restarts the state machine
statemachine.rx_fifo: Return the number of RX FIFO items. 0 if empty
statemachine.tx_fifo: Return the number of TX FIFO items. 0 if empty

restart() seems to be the most useful one, as it resets the state
machine to the initial state without the need to re-instantiation.
It also makes PIO code easier, because then stalling as error
state can be unlocked.

rx_fifo() is also useful, for MP code to check for data & timeout if
no data arrived. Complex logic is easier handled in Python code than
in PIO code.

tx_fifo() can be useful to check states where data is not
processed. And is mostly for symmetry.

These are tiny functions, with 2 or 3 active codelines only plus the
definition of their names.