rp2_pio.c: Fix sm.get(buf) size expectation.
sm.get(buf) was waiting for another item more than the length of
the supplied buffer. Even if this item was not stored, sm_get stuck
in trying to get an item from the RX fifo.
As part of that, the edge case fo a zero length buffer was moved up
to the section, where the function arguments are handled. In case
of a zero length buffer, sm.get() now returns immediately that buffer.
It could as well return None.