build.yml: Fix 'Print failure info' to not error
Print Failure Info is for printing differences detected by 'Test all'.  When some other step fails, then "*.exp" doesn't match any files at all, and _this_ step fails too.

We will assume that Github Actions always runs with bash as the shell (as it does today). In this case, we can set the  "nullglob" shell option, so that a non-existent glob expands to nothing, instead of to itself.