machine_uart.c: Add buffered transfer of data.
Instantiation and init now support the rxbuf and txbuf keywords for
setting the buffer size. The default and minimal size is 256.
The maximal size is 32766.

uart.write() still includes checks for timeout, even if it is very
unlikely to happen due to a) lack of flow control support and
b) the minimal timeout values being longer than the time it needs
to send a byte .