rp2:moduos.c, main.c: Add uos.urandom().
The implementation samples rosc.randombits at a frequency lower
than the oscillator frequency. The gives better random values.
In addition, for a 8 bit value 32 samples are taken and fed through
a 8 bit CRC, distributing the oversampling over the byte.
The resulting sampling rate is about 30k/sec.

The RNG does not include testing of error conditions, like the ROSC
being in sync with the sampling or completely failing. Making the
interim value static causes it to perform a little bit better in short
sync or drop-out situations.

The output of uos.urandom() performs well with the NIST800-22 test
suite. In my trial it passed all tests of the sts 2.1.2 test suite.