extmod/modonewire.c: Fix onewire timing and wait for clear bus.
The timing of the onewire module was way too fast when reading.
The write timing was correct, with a 10us low pulse for 1 and
60us pulse for 0. But read would pulse low for 5us and then sample
the wire after 10us. This is way too fast since a 1 can be
a pulse of up to 15us and would be read as 0. Plus the whole bit
time for read was only 50us, while a client may output a pulse
of up to 60us to indicate a 0. Compounding the problem, the code
would not check if the bus was high before starting a bit pulse,
just blindly plowing through whatever the slave might have been
sending with the next bit before the slave was done.

This code now outputs a 10us pulse for read to be consistent with
the write, then samples the bus after 30us as is the case with
most implementations I found. The total read bit time is 60us.
The code also first checks whether the bus is high before starting
a bit, and waits in 10us increments up to 100us before starting to
output a bit pulse.