extmod/re1.5: Error in _compilecode if value out of range.
.. adding a test for this case.

Locally, this change increases the size of PYBV10 by +80 bytes text.

The complication where the argument is checked for constness reduces
code size slightly (-14 bytes).  This gcc extension is supported by clang
and is already used in the stm and teensy ports.

This change assumes that the results that overflow an int8_t do not
overflow a platform int.

Closes: #7078
Signed-off-by: Jeff Epler <jepler@gmail.com>