rp2: Import uarray instead of array.
Some forum users noticed that `sm.exec()` took longer the more was
present on the flash filesystem connected to the RP2040.

They traced this back to the `array` import inside `asm_pio()`, which is
causing MicroPython to scan the filesystem.

uarray is a built-in module, so importing it shouldn't require scanning
the filesystem.

We avoid moving the import to the top-level in order to keep the
namespace clean; we don't want to accidentally expose `rp2.array`.