unix: Improve command line argument processing.
Per CPython everything which comes after the command, module or file
argument is not an option for the interpreter itself.  Hence the
processing of options should stop when encountering those, and the
remainder be passed as sys.argv.  Note he latter was already the
case for a module or file but not for a command.

This fixes issues like 'micropython myfile.py -h' showing the help
and exiting instead of passing '-h' as sys.argv[1], likewise for
'-X <something>' being treated as a special option no matter where it
occurs on the command line.