machine_uart.c: Move handling of lock flags.
Setting and releasing the flags is moved into the API functions instaed
of just passing an argument. That shrinks the ISR code.
The write-lock is still present, event if I could not yet create a test
case of failing without the lock. The check for timeout on write is
also still included, even if it cannot happen due to the absence of
a transmit flow control.