mimxrt/machine_pin: Implemented mode method.
Implemented mode method and upgraded make-pins.py and
related files to not treat machine_pin objects as constanst.
Unfortunately reading out the current configuration of a PAD is
relatively complex. There is no function provided by the NXP HAL
afaik and therefore it is easier to implement a field in the
machine_pin object which contains that information. The downside
of that is increased RAM usage.