mimxrt/machine_pin: Added init ad drive methods.
Added option to configure the drive strength of pins
(used for impedance matching). Since i.MXRT supports more options
there than other ports in addition to the known LOW, MED, and HIGH
intermediate levels are available. On top of that the output driver
can even be disabled for inputs for instance.