mimxrt: Provide a RTC module.
Initial version, using the LP rtc clock.
It provides setting the Date & Time with rtc.init() or rtc.datetime()
and reading date & time with rtc.datetime() or rtc.now().

If not battery is supplied, the default boot time/date is 1970/1/1 0:0:0.
With a battery, the clock continues to run even when the board is not
powered. The clock is quite precise. If not, using rtc.calibration()
may help.

Possible extension: set an alarm with alarm handler. For short periods
< 70 minutes this is also possible with the timer module,
for longer periods this may have to be done.