mimxrt: Return to common MP format of machine.RTC.
This change moves the datetime format back to the on used by all
other MP ports:

Year, Month, Day, Weekday, Hour, Minute, Second, Microsecond.

Weeday is a number between 0 and 6, with 0 assigned to Monday.
It has to be provided when setting the RTC with datetime(), but
will be ignored on entry and calculated when needed.

The weekday method was removed, since that is now again a part of the
datetime tuple.