ports/unix/modffi.c: FFI fixes.
Fixes and improvements to ffi callback

Added an optional 'lock' kwarg to callback that locks gc and scheduler.
This allows the callback to be invoked asynchronously in 'interrupt
context', for example as a signal handler.

Added 'cfun' member function to callback, that allows retrieving the C
callback function address.
This is needed when the callback should be set to a struct field.

Signed-off-by: Amir Gonnen <amirgonnen@gmail.com>