ports/unix/modffi.c: FFI fixes.
Fixes and improvements to ffi callabck

Add an optional 'lock' kwarg to callback that locks gc and scheduler.
This allows the callback to be invoked asynchronously in 'interrupt
context', for example as a signal handler

Added buffer protocol to callback that allows retrieving the C callback
function. This is needed when the callback should be set to a struct
field.

Signed-off-by: Amir Gonnen <amirgonnen@gmail.com>