ports/rp2: Use 0=Monday datetime convention in RTC.
The RTC in rp2 can store any, even wrong, number as a weekday in RTC. It
was, however, discussed in #7394 that we would like to unify all ports
and use 0 as Monday, not Sunday in machine.RTC implementation.

This patch makes sure that the default date set in RTC is adheres to
this convention. It also fixes the example in quickref to use proper
weekday to avoid confusion.

Signed-off-by: Krzysztof Adamski <k@japko.eu>