rp2/machine_rtc: RTC compute weekday autmatically.
Calculating the weekday each time you want to set a date is error prone
and tiresome. Micropython can do it on its own - hardware on some ports
do not support storing weekday in hardware and always computes it on the
fly, ignoring the value given to the constructor.

During discussion for #7432 PR, we have concluded, that there seems to
be no obvious reason to let user set the weekday to incorrect value so
it makes sense to just ignore the provided weekday value and always
compute the correct value.  This patch introduces this change for rp2
port.

Signed-off-by: Krzysztof Adamski <k@japko.eu>