top: Simplify buffer protocol to just a "get buffer" callback.
The buffer protocol type only has a single member, and this
layout creates problems for the upcoming split/slot-index
mp_obj_type_t layout optimisations.

If we need to make the buffer protocol more sophisticated
in the future either we can rely on the mp_obj_type_t
optimisations to just add additional slots to mp_obj_type_t
or re-visit the buffer protocol then.

This change is a no-op in terms of generated code.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>