mimxrt: Support PWM using the FLEXPWM and QTMR modules.
Frequency range 15Hz/18Hz to > 1 MHz, with decreasing resolution of
the duty cycle. The basic API is supported as documentated, except
that keyword parameters are accepted for both the instatiaton and
the PWM.init() call.

Extensions:
- Support PWM for channel pairs.

Channel pairs are declared by supplying 2-element tuples for the
pins. The two channels of a pair must be the A/B channel of a
FLEXPWM module. These form than a complementary pair.

Additional supported keywprd parameters:

center=value Defines the center position of a pulse within the
  pulse cycle. The align keyword is actually shortcut for center.

sync=True|False: If set to True, the channels will be synchronized to
  a submodule 0 channel, which has already to be enabled.

align=PWM.MIDDLE | PMW.BEGIN | PWM.END. It defines,
  whether synchronized channels are Center-Aligned or Edge-
  aligned. The channels must be either complementary a channel pair or
  a group of synchronized channels. It may as well be applied to a
  single channel, but withiout any benefit.

invert= 0..3. Controls ouput inversion of the pins. Bit
  0 controls the first pin, bit 1 the second.

start=True|False start=False causes the PWM output to not start
  immediately. The pwm.start() method can later be used to start a
  group of channels at the same time.

deadtime=time_ns time of complementary channels for delaying the rising
  slope.

xor=0|1|2 xor causes the output of channel A and B to be xored.
  If applied to a    X channel, it shows the value oif A ^ B.
  If applied to an A or B channel, both channel show the xored signal
  for xor=1. For xor=2, the xored signal is split between channels A
  and B. See also the Reference Manual, chapter about double pulses.
  The behavior of xor=2 can also be achieved using the center method
  for locating a pulse within a clock period.

The output is enabled for board pins only.

CPU pins may still be used for FLEXPWM, e.g. as sync source, but the
signal will not be routed to the output. That applies only to FLEXPWM
pins. The use of QTMR pins which are not board pins will be rejected.

Recent activities:

- Implement the review comments.
- Fix PWM glitch for FlexPWM and QTMR
  The FLEXPWM glitch was caused by the function ticks_delay_us64(),
  which is called by sleep(), and sleep() was used in the test script.
  The cure consisted of removing a _WFE() statement, which caused
  already earlier trouble.