check vid/pid: find topdir relative to this script
@dhalbert noticed that if a fork was called `micropython`, the
existing check could fail during CI because it will be cloned into a
directory also called `micropython`.

Instead of hardcoding a range of strings that are OK as top directories,
find the location of the top directory relative to the script.

This will be conflicty after #5069 is merged, but it should be easy to
fix.